\name{ggsports-package}
\alias{ggsports-package}
\alias{ggsports}
\docType{package}
\title{
\packageTitle{ggsports}
}
\description{
\packageDescription{ggsports}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ggsports}
\packageIndices{ggsports}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ggsports}

Maintainer: \packageMaintainer{ggsports}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
